api/requirements.txt
flask
flask-restful
psycopg2-binary==2.9.1
Flask-Cors==3.0.10